This would create a new player class and shelving scheme that would preserve player data (name, single session high scores, and cumulative wallet) between sessions.
This would also create new functions that would update single session scores and wallets, and display prior player history data at the conclusion of each session.  This display would use '\t' as a way to space out text in this display, as opposed to 'LINE_OF_SPACES' that would still be employed by the hand display functions (this might be subject to use of '\t' in a further update).

This would create logic to detect blackjack (initial hands containing a Jack and an Ace) in both the player's hand and the dealer's hands, and the ensuing consequences (instant win, loss, or push). This would employ the logic by wrapping most of the linear code into the 'else' portion of an 'if' statement.