This commit would:
Add language to import a new module (still under construction) for splitting hands.  The option to split hands (and all of the code enabling this option) would be provided in a future update.

Modify the function which serves to display the history of players and select statistics for those players.  The function would now sort the data to display player history ordered by winning percentage.

Redefine "blackjack" hand to mean a combination of a 10-point value card and an Ace (instead of a Jack and an Ace).

Give players a 50% bonus to winning blackjack hands (21 point initial hands).

Wrap monetary variables in an 'int' call, to yield whole numbers in the user displays.

Create double double feature.

Create surrender feature, and initialize 'surrender' variable to None.  ('surrender' variable is switched to 'True' if it is triggered during a hand).

Move logic for detecting an empty wallet and auto-exiting the game into a function call.

Modify the decision table (e.g, 'Hit', 'Stand', 'Double down', 'Surrender', etc.)  to display options one at a time, per line.